avg->float

+++ HttpServer/express/service/commonUtils.js
@@ -103,7 +103,7 @@ function avg(arrays) {
   let len = arrays.length;
   let result = 0;
   for (let i = 0; i < len; i++) {
-    result += parseInt(arrays[i]);
+    result += parseFloat(arrays[i]);
   }
 
   return (result / len).toFixed(2);
